tests/wpt/web-platform-tests/referrer-policy/css-integration/child-css/external-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - Child css from external stylesheet</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/referrer-policy/generic/common.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that child css are loaded with the referrer and referrer policy
    from the external stylesheet.</p>

    <div class="styled"></div>

    <script>
      async_test(function(css_test) {
        let id = token();
        let url_prefix = location.protocol + "//www1." + location.hostname + ":" + location.port;
        let css_url = url_prefix +
          "/referrer-policy/generic/subresource/stylesheet.py?id=" + id +
          "&import-rule" + "&referrer-policy=no-referrer";
        let check_url = url_prefix + "/referrer-policy/generic/subresource/stylesheet.py" +
                        "?id=" + id + "&report-headers";

        let link = document.createElement("link");
        link.href = css_url;
        link.rel = "stylesheet";
        link.onload = function() {
          css_test.step_timeout(function() {
              queryXhr(check_url, function(message) {
                  assert_own_property(message, "headers");
                  assert_equals(message.referrer, undefined);
                  css_test.done();
              }, null, null, css_test);
          }, 1000);
        };
        document.head.appendChild(link);
      }, "Child css from external stylesheet.");
    </script>

    <div id="log"></div>
  </body>
</html>
